+++ src/jpl_cad_ollama_explorer/__init__.py
@@ -1,2 +1,2 @@
 __app_name__ = "JPL CAD Ollama Explorer"
-__version__ = "0.1.27"
+__version__ = "0.1.31"